requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
SQLAlchemy==2.0.28
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
pytest==8.1.1
httpx==0.27.0
pytest-asyncio==0.23.6